move snapshot api to storage group
