+++ hack/.golint_failures
@@ -296,10 +296,6 @@ pkg/registry/core/service/portallocator
 pkg/registry/core/service/portallocator/controller
 pkg/registry/core/service/storage
 pkg/registry/core/serviceaccount/storage
-pkg/registry/core/volumesnapshot
-pkg/registry/core/volumesnapshot/storage
-pkg/registry/core/volumesnapshotdata
-pkg/registry/core/volumesnapshotdata/storage
 pkg/registry/events/rest
 pkg/registry/extensions/controller/storage
 pkg/registry/extensions/ingress
@@ -330,6 +326,10 @@ pkg/registry/settings/rest
 pkg/registry/storage/rest
 pkg/registry/storage/storageclass
 pkg/registry/storage/storageclass/storage
+pkg/registry/storage/volumesnapshot
+pkg/registry/storage/volumesnapshot/storage
+pkg/registry/storage/volumesnapshotdata
+pkg/registry/storage/volumesnapshotdata/storage
 pkg/routes
 pkg/security/apparmor
 pkg/security/podsecuritypolicy
