+++ pkg/apis/core/register.go
@@ -93,10 +93,6 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&RangeAllocation{},
 		&ConfigMap{},
 		&ConfigMapList{},
-		&VolumeSnapshot{},
-		&VolumeSnapshotList{},
-		&VolumeSnapshotData{},
-		&VolumeSnapshotDataList{},
 	)
 
 	return nil
