+++ pkg/apis/core/validation/validation.go
@@ -1881,142 +1881,6 @@ func ValidatePersistentVolumeClaimUpdate(newPvc, oldPvc *core.PersistentVolumeCl
 	return allErrs
 }
 
-// ValidateVolumeSnapshotName checks that a name is appropriate for a
-// VolumeSnapshot object.
-var ValidateVolumeSnapshotName = NameIsDNSSubdomain
-
-func ValidateVolumeSnapshot(vs *core.VolumeSnapshot) field.ErrorList {
-	metaPath := field.NewPath("metadata")
-	allErrs := ValidateObjectMeta(&vs.ObjectMeta, true, ValidateVolumeSnapshotName, metaPath)
-	specPath := field.NewPath("spec")
-
-	if len(vs.Spec.PersistentVolumeClaimName) == 0 {
-		allErrs = append(allErrs, field.Required(specPath.Child("persistentVolumeClaimName"), ""))
-	} else {
-		for _, msg := range ValidatePersistentVolumeName(vs.Spec.PersistentVolumeClaimName, false) {
-			allErrs = append(allErrs, field.Invalid(specPath.Child("persistentVolumeClaimName"), vs.Spec.PersistentVolumeClaimName, msg))
-		}
-	}
-	return allErrs
-}
-
-// ValidateVolumeSnapshotUpdate tests to see if the update is legal for an end user to make.
-// newVs is updated with fields that cannot be changed.
-func ValidateVolumeSnapshotUpdate(newVs, oldVs *core.VolumeSnapshot) field.ErrorList {
-	allErrs := field.ErrorList{}
-	allErrs = ValidateVolumeSnapshot(newVs)
-
-	// PersistentVolumeClaimName should be immutable after creation.
-	if newVs.Spec.PersistentVolumeClaimName != oldVs.Spec.PersistentVolumeClaimName {
-		allErrs = append(allErrs, field.Forbidden(field.NewPath("spec", "persistentvolumeclaimname"), "is immutable after creation"))
-	}
-	return allErrs
-}
-
-// ValidateVolumeSnapshotStatusUpdate validates an update to status of a VolumeSnapshot.
-func ValidateVolumeSnapshotStatusUpdate(newVs, oldVs *core.VolumeSnapshot) field.ErrorList {
-	allErrs := ValidateObjectMetaUpdate(&newVs.ObjectMeta, &oldVs.ObjectMeta, field.NewPath("metadata"))
-	if len(newVs.ResourceVersion) == 0 {
-		allErrs = append(allErrs, field.Required(field.NewPath("resourceVersion"), ""))
-	}
-	if len(newVs.Spec.PersistentVolumeClaimName) == 0 {
-		allErrs = append(allErrs, field.Required(field.NewPath("Spec", "persistentVolumeClaimName"), ""))
-	}
-	return allErrs
-}
-
-// ValidateVolumeSnapshotDataName checks that a name is appropriate for a
-// VolumeSnapshotData object.
-var ValidateVolumeSnapshotDataName = NameIsDNSSubdomain
-
-func ValidateVolumeSnapshotData(vsd *core.VolumeSnapshotData) field.ErrorList {
-	metaPath := field.NewPath("metadata")
-	allErrs := ValidateObjectMeta(&vsd.ObjectMeta, false, ValidateVolumeSnapshotDataName, metaPath)
-
-	specPath := field.NewPath("spec")
-
-	numVolumeSnapshotDataSource := 0
-	if vsd.Spec.HostPath != nil {
-		if numVolumeSnapshotDataSource > 0 {
-			allErrs = append(allErrs, field.Forbidden(specPath.Child("hostPath"), "may not specify more than 1 volume snapshot data source"))
-		} else {
-			numVolumeSnapshotDataSource++
-			if len(vsd.Spec.HostPath.Path) == 0 {
-				allErrs = append(allErrs, field.Required(specPath.Child("hostPath", "path"), ""))
-			}
-		}
-	}
-	if vsd.Spec.AWSElasticBlockStore != nil {
-		if numVolumeSnapshotDataSource > 0 {
-			allErrs = append(allErrs, field.Forbidden(specPath.Child("AWSElasticBlockStore"), "may not specify more than 1 volume snapshot data source"))
-		} else {
-			numVolumeSnapshotDataSource++
-			if len(vsd.Spec.AWSElasticBlockStore.SnapshotID) == 0 {
-				allErrs = append(allErrs, field.Required(specPath.Child("AWSElasticBlockStore", "SnapshotID"), ""))
-			}
-		}
-	}
-	if vsd.Spec.GlusterSnapshotVolume != nil {
-		if numVolumeSnapshotDataSource > 0 {
-			allErrs = append(allErrs, field.Forbidden(specPath.Child("GlusterSnapshotVolume"), "may not specify more than 1 volume snapshot data source"))
-		} else {
-			numVolumeSnapshotDataSource++
-			if len(vsd.Spec.GlusterSnapshotVolume.SnapshotID) == 0 {
-				allErrs = append(allErrs, field.Required(specPath.Child("GlusterSnapshotVolume", "SnapshotID"), ""))
-			}
-		}
-	}
-	if vsd.Spec.CinderSnapshot != nil {
-		if numVolumeSnapshotDataSource > 0 {
-			allErrs = append(allErrs, field.Forbidden(specPath.Child("CinderSnapshot"), "may not specify more than 1 volume snapshot data source"))
-		} else {
-			numVolumeSnapshotDataSource++
-			if len(vsd.Spec.CinderSnapshot.SnapshotID) == 0 {
-				allErrs = append(allErrs, field.Required(specPath.Child("CinderSnapshot", "SnapshotID"), ""))
-			}
-		}
-	}
-	if vsd.Spec.GCEPersistentDiskSnapshot != nil {
-		if numVolumeSnapshotDataSource > 0 {
-			allErrs = append(allErrs, field.Forbidden(specPath.Child("GCEPersistentDiskSnapshot"), "may not specify more than 1 volume snapshot data source"))
-		} else {
-			numVolumeSnapshotDataSource++
-			if len(vsd.Spec.GCEPersistentDiskSnapshot.SnapshotName) == 0 {
-				allErrs = append(allErrs, field.Required(specPath.Child("GCEPersistentDiskSnapshot", "SnapshotName"), ""))
-			}
-		}
-	}
-	if numVolumeSnapshotDataSource == 0 {
-		allErrs = append(allErrs, field.Required(specPath, "must specify a volume snapshot data source"))
-	}
-
-	return allErrs
-}
-
-// ValidateVolumeSnapshotDataUpdate tests to see if the update is legal for an end user to make.
-// newVsd is updated with fields that cannot be changed.
-func ValidateVolumeSnapshotDataUpdate(newVsd, oldVsd *core.VolumeSnapshotData) field.ErrorList {
-	allErrs := field.ErrorList{}
-	allErrs = ValidateVolumeSnapshotData(newVsd)
-
-	// PersistentVolumeSource should be immutable after creation.
-	if !apiequality.Semantic.DeepEqual(newVsd.Spec.VolumeSnapshotDataSource, oldVsd.Spec.VolumeSnapshotDataSource) {
-		allErrs = append(allErrs, field.Forbidden(field.NewPath("spec", "volumesnapshotdatasource"), "is immutable after creation"))
-	}
-
-	return allErrs
-}
-
-// ValidateVolumeSnapshotDataStatusUpdate tests to see if the status update is legal for an end user to make.
-// newVsd is updated with fields that cannot be changed.
-func ValidateVolumeSnapshotDataStatusUpdate(newVsd, oldVsd *core.VolumeSnapshotData) field.ErrorList {
-	allErrs := ValidateObjectMetaUpdate(&newVsd.ObjectMeta, &oldVsd.ObjectMeta, field.NewPath("metadata"))
-	if len(newVsd.ResourceVersion) == 0 {
-		allErrs = append(allErrs, field.Required(field.NewPath("resourceVersion"), ""))
-	}
-	return allErrs
-}
-
 // Provide an upgrade path from PVC with storage class specified in beta
 // annotation to storage class specified in attribute. We allow update of
 // StorageClassName only if following four conditions are met at the same time:
