+++ staging/src/k8s.io/api/core/v1/register.go
@@ -89,10 +89,6 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&RangeAllocation{},
 		&ConfigMap{},
 		&ConfigMapList{},
-		&VolumeSnapshot{},
-		&VolumeSnapshotList{},
-		&VolumeSnapshotData{},
-		&VolumeSnapshotDataList{},
 	)
 
 	// Add common types
