+++ staging/src/k8s.io/api/storage/v1alpha1/register.go
@@ -43,6 +43,10 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 	scheme.AddKnownTypes(SchemeGroupVersion,
 		&VolumeAttachment{},
 		&VolumeAttachmentList{},
+		&VolumeSnapshot{},
+		&VolumeSnapshotList{},
+		&VolumeSnapshotData{},
+		&VolumeSnapshotDataList{},
 	)
 
 	metav1.AddToGroupVersion(scheme, SchemeGroupVersion)
